+++ common/src/main/java/io/github/mortuusars/horseman/mixin/HorsemanMixinPlugin.java
@@ -15,7 +15,7 @@ import java.util.function.Supplier;
 public class HorsemanMixinPlugin implements IMixinConfigPlugin {
 
     private static final Map<String, Supplier<Boolean>> CONDITIONS = ImmutableMap.of(
-            "io.github.mortuusars.horseman.mixin.ServerGamePacketListenerImplMixin",
+            "io.github.mortuusars.horseman.mixin.FixMovedWrongly_ServerGamePacketListenerImplMixin",
                 () -> !PlatformHelper.isModLoading("horsebuff")
     );